Possibility to add to the link CRU page with preformatted payload:
The autoformatting behaviour of RawFileWriter::addData can be changed
to preformatted page storage by adding an extra true argument in the end of addData method, e.g.
writer.addData(rdh, ir, gsl::span( (char*)payload_f, payload_f_size ), true );

In the case provided data blob is interpretted as a fully formatted CRU page payload (i.e. it lacks
the RDH which will be added by the writer) of the maximum size 8192-sizeof(RDH) = 8128 bytes.
The writer will create a new CRU page with provided payload, equipping it with the proper RDH:
copying already stored RDH of the current HBF, if the interaction record  belongs to the same HBF
or generating new RDH for new HBF otherwise (and filling all missing HBFs in-between).
In case the payload size exceeds maximum, an exception will be thrown w/o any attempt to split the page.

+++ Detectors/Raw/include/DetectorsRaw/HBFUtils.h
@@ -127,7 +127,7 @@ struct HBFUtils : public o2::conf::ConfigurableParamHelper<HBFUtils> {
 
   // some fields of the same meaning have different names in the RDH of different versions
   static uint32_t getHBOrbit(const void* rdhP);
-  static uint32_t getHBBC(const void* rdhP);
+  static uint16_t getHBBC(const void* rdhP);
   static IR getHBIR(const void* rdhP);
 
   static void printRDH(const void* rdhP);
@@ -138,8 +138,8 @@ struct HBFUtils : public o2::conf::ConfigurableParamHelper<HBFUtils> {
   static uint32_t getHBOrbit(const o2::header::RAWDataHeaderV4& rdh) { return rdh.heartbeatOrbit; }
   static uint32_t getHBOrbit(const o2::header::RAWDataHeaderV5& rdh) { return rdh.orbit; }
 
-  static uint32_t getHBBC(const o2::header::RAWDataHeaderV4& rdh) { return rdh.heartbeatBC; }
-  static uint32_t getHBBC(const o2::header::RAWDataHeaderV5& rdh) { return rdh.bunchCrossing; }
+  static uint16_t getHBBC(const o2::header::RAWDataHeaderV4& rdh) { return rdh.heartbeatBC; }
+  static uint16_t getHBBC(const o2::header::RAWDataHeaderV5& rdh) { return rdh.bunchCrossing; }
 
   static IR getHBIR(const o2::header::RAWDataHeaderV4& rdh) { return {uint16_t(rdh.heartbeatBC), uint32_t(rdh.heartbeatOrbit)}; }
   static IR getHBIR(const o2::header::RAWDataHeaderV5& rdh) { return {uint16_t(rdh.bunchCrossing), uint32_t(rdh.orbit)}; }
